+++ rotkehlchen/chain/arbitrum_one/modules/monerium/decoder.py
@@ -7,6 +7,7 @@ from .constants import ARBITRUM_MONERIUM_ADDRESSES
 if TYPE_CHECKING:
     from rotkehlchen.chain.arbitrum_one.node_inquirer import ArbitrumOneInquirer
     from rotkehlchen.chain.evm.decoding.base import BaseEvmDecoderTools
+    from rotkehlchen.externalapis.monerium import Monerium
     from rotkehlchen.user_messages import MessagesAggregator
 
 
@@ -17,10 +18,12 @@ class MoneriumDecoder(MoneriumCommonDecoder):
             arbitrum_one_inquirer: 'ArbitrumOneInquirer',
             base_tools: 'BaseEvmDecoderTools',
             msg_aggregator: 'MessagesAggregator',
+            monerium_api: 'Monerium | None',
     ) -> None:
         super().__init__(
             evm_inquirer=arbitrum_one_inquirer,
             base_tools=base_tools,
             msg_aggregator=msg_aggregator,
+            monerium_api=monerium_api,
             monerium_token_addresses=ARBITRUM_MONERIUM_ADDRESSES,
         )