+++ rotkehlchen/chain/gnosis/manager.py
@@ -9,6 +9,7 @@ from .tokens import GnosisTokens
 from .transactions import GnosisTransactions
 
 if TYPE_CHECKING:
+    from rotkehlchen.externalapis.monerium import Monerium
     from rotkehlchen.premium.premium import Premium
 
     from .node_inquirer import GnosisInquirer
@@ -16,7 +17,12 @@ if TYPE_CHECKING:
 
 class GnosisManager(EvmManager, CurveManagerMixin):
 
-    def __init__(self, node_inquirer: 'GnosisInquirer', premium: 'Premium | None' = None) -> None:
+    def __init__(
+            self,
+            node_inquirer: 'GnosisInquirer',
+            premium: 'Premium | None' = None,
+            monerium: 'Monerium | None' = None,
+    ) -> None:
         transactions = GnosisTransactions(
             gnosis_inquirer=node_inquirer,
             database=node_inquirer.database,
@@ -37,6 +43,7 @@ class GnosisManager(EvmManager, CurveManagerMixin):
                 gnosis_inquirer=node_inquirer,
                 transactions=transactions,
                 premium=premium,
+                monerium=monerium,
             ),
             accounting_aggregator=GnosisAccountingAggregator(
                 node_inquirer=node_inquirer,