+++ rotkehlchen/chain/polygon_pos/decoding/decoder.py
@@ -13,6 +13,7 @@ if TYPE_CHECKING:
     from rotkehlchen.chain.polygon_pos.node_inquirer import PolygonPOSInquirer
     from rotkehlchen.chain.polygon_pos.transactions import PolygonPOSTransactions
     from rotkehlchen.db.dbhandler import DBHandler
+    from rotkehlchen.externalapis.monerium import Monerium
     from rotkehlchen.premium.premium import Premium
 
 logger = logging.getLogger(__name__)
@@ -29,6 +30,7 @@ class PolygonPOSTransactionDecoder(EVMTransactionDecoder):
             polygon_pos_inquirer: 'PolygonPOSInquirer',
             transactions: 'PolygonPOSTransactions',
             premium: 'Premium | None' = None,
+            monerium: 'Monerium | None' = None,
     ):
         super().__init__(
             database=database,
@@ -45,6 +47,7 @@ class PolygonPOSTransactionDecoder(EVMTransactionDecoder):
                 exceptions_mappings=V1_TO_V2_MONERIUM_MAPPINGS,
             ),
             premium=premium,
+            monerium=monerium,
             addresses_exceptions=dict.fromkeys(POLYGON_MONERIUM_LEGACY_ADDRESSES, MONERIUM_V2_CONTRACTS_BLOCK),  # noqa: E501
         )